IA/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
tensorflow==2.18.0
numpy==1.26.0
opencv-python==4.8.1.78
mtcnn==0.1.1
pillow==10.0.1